+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Scripts.hs
@@ -31,6 +31,7 @@ module Cardano.Ledger.Alonzo.Scripts (
   validScript,
   transProtocolVersion,
   eqAlonzoScriptRaw,
+  translateAlonzoScript,
 
   -- * Cost Model
   CostModel,
@@ -54,7 +55,7 @@ module Cardano.Ledger.Alonzo.Scripts (
 )
 where
 
-import Cardano.Ledger.Allegra.Scripts (Timelock, eqTimelockRaw)
+import Cardano.Ledger.Allegra.Scripts (Timelock, eqTimelockRaw, translateTimelock)
 import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.BaseTypes (
   BoundedRational (unboundRational),
@@ -173,6 +174,14 @@ data AlonzoScript era
   | PlutusScript !Plutus
   deriving (Eq, Generic, NoThunks)
 
+translateAlonzoScript ::
+  (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) =>
+  AlonzoScript era1 ->
+  AlonzoScript era2
+translateAlonzoScript = \case
+  TimelockScript ts -> TimelockScript $ translateTimelock ts
+  PlutusScript ps -> PlutusScript ps
+
 instance (EraScript era, Script era ~ AlonzoScript era) => Show (AlonzoScript era) where
   show (TimelockScript x) = "TimelockScript " ++ show x
   show s@(PlutusScript plutus) =
@@ -195,9 +204,13 @@ isPlutusScript (TimelockScript _) = False
 
 instance Crypto c => EraScript (AlonzoEra c) where
   type Script (AlonzoEra c) = AlonzoScript (AlonzoEra c)
+
+  upgradeScript = TimelockScript . translateTimelock
+
   phaseScript PhaseOneRep (TimelockScript s) = Just (Phase1Script s)
   phaseScript PhaseTwoRep (PlutusScript plutus) = Just (Phase2Script plutus)
   phaseScript _ _ = Nothing
+
   scriptPrefixTag script =
     case script of
       TimelockScript _ -> nativeMultiSigTag -- "\x00"