+++ eras/mary/impl/src/Cardano/Ledger/Mary/Scripts.hs
@@ -1,20 +1,10 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE DerivingVia #-}
-{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Mary.Scripts (
@@ -22,7 +12,7 @@ module Cardano.Ledger.Mary.Scripts (
 )
 where
 
-import Cardano.Ledger.Allegra.Scripts
+import Cardano.Ledger.Allegra.Scripts (Timelock, translateTimelock)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Mary.Era (MaryEra)
@@ -35,6 +25,10 @@ type instance SomeScript 'PhaseOne (MaryEra c) = Timelock (MaryEra c)
 -- for the ValidateScript instance in MultiSig
 instance Crypto c => EraScript (MaryEra c) where
   type Script (MaryEra c) = Timelock (MaryEra c)
+
+  upgradeScript = translateTimelock
+
   scriptPrefixTag _script = nativeMultiSigTag -- "\x00"
+
   phaseScript PhaseOneRep timelock = Just (Phase1Script timelock)
   phaseScript PhaseTwoRep _ = Nothing