+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/AlonzoBBODY.hs
@@ -662,7 +662,7 @@ testBBodyState pf =
       poolID = hashKey . vKey . coerceKeyRole $ coldKeys
       example1UtxoSt = smartUTxOState (pp pf) utxo totalDeposits (Coin 40) def
       -- the default CertState 'def' means that the 'totalDeposits' must be 0
-      totalDeposits = (Coin 0)
+      totalDeposits = Coin 0
    in BbodyState
         (LedgerState example1UtxoSt def)
         (BlocksMade $ Map.singleton poolID 1)